# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación para API REST de Recetas",
        "contact": {
            "email": "eescobar@cdhidalgo.tecnm.mx"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apibecas2-production.up.railway.app"
        }
    ],
    "paths": {
...
        "/api/estudiantes": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtener lista de estudiantes",
                "operationId": "e39ad3a168705ec612b7f3937faaaa1c",
                "responses": {
                    "200": {
                        "description": "Lista de estudiantes recuperada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "numero_control": {
                                                "type": "string",
                                                "example": "19141142"
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Juan"
                                            },
                                            "apellido_paterno": {
                                                "type": "string",
                                                "example": "Pérez"
                                            },
                                            "apellido_materno": {
                                                "type": "string",
                                                "example": "García"
                                            },
                                            "id_carrera": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "carrera": {
                                                "type": "string",
                                                "example": "Ingeniería en Sistemas"
                                            },
                                            "correo": {
                                                "type": "string",
                                                "example": "juan.perez@ejemplo.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'Ver registros'"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Crear un nuevo estudiante",
                "operationId": "ef4e853a87acd60ea86b0e5579eb5208",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero_control": {
                                        "type": "string",
                                        "example": "19141142"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "id_carrera": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez@ejemplo.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudiante creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "numero_control": {
                                            "type": "string",
                                            "example": "19141142"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Juan"
                                        },
                                        "apellido_paterno": {
                                            "type": "string",
                                            "example": "Pérez"
                                        },
                                        "apellido_materno": {
                                            "type": "string",
                                            "example": "García"
                                        },
                                        "id_carrera": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "correo": {
                                            "type": "string",
                                            "example": "juan.perez@ejemplo.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'CrearEliminar'"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/estudiantes/{id}": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtener un estudiante específico",
                "operationId": "fc77eb37ffbc9deb381d9e78cf44a6ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "numero_control": {
                                            "type": "string",
                                            "example": "19141142"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Juan"
                                        },
                                        "apellido_paterno": {
                                            "type": "string",
                                            "example": "Pérez"
                                        },
                                        "apellido_materno": {
                                            "type": "string",
                                            "example": "García"
                                        },
                                        "id_carrera": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "carrera": {
                                            "type": "string",
                                            "example": "Ingeniería en Sistemas"
                                        },
                                        "correo": {
                                            "type": "string",
                                            "example": "juan.perez@ejemplo.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'Ver registros'"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Actualizar un estudiante existente",
                "operationId": "c3348f934138e9b2983eebcb6ed6b537",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Carlos"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "id_carrera": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez.nuevo@ejemplo.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estudiante actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "numero_control": {
                                            "type": "string",
                                            "example": "19141142"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Juan Carlos"
                                        },
                                        "apellido_paterno": {
                                            "type": "string",
                                            "example": "Pérez"
                                        },
                                        "apellido_materno": {
                                            "type": "string",
                                            "example": "García"
                                        },
                                        "id_carrera": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "correo": {
                                            "type": "string",
                                            "example": "juan.perez.nuevo@ejemplo.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'Modificar registros'"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Eliminar un estudiante",
                "operationId": "a1633169ea7c0b3a90e60fa98cb7dff0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'CrearEliminar'"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum API token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}